requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.21.0
python-multipart==0.0.18
pyyaml==6.0.2
Pillow==11.1.0
